Refactor: Consolidate component imports and remove unused files in components-tab [ED-20800]
