+++ packages/packages/core/editor-components/src/init.ts
@@ -2,7 +2,7 @@ import { injectIntoTop } from '@elementor/editor';
 import { injectTab } from '@elementor/editor-elements-panel';
 import { __ } from '@wordpress/i18n';
 
-import { Components } from './components/components-tab/components';
+import { Components } from './components/components';
 import { CreateComponentForm } from './components/create-component-form/create-component-form';
 
 export function init() {
